+++ Matchismo2013Fall/Model/Card.h
@@ -16,6 +16,6 @@
 @property (nonatomic, getter = isChosen) BOOL chosen;
 @property (nonatomic, getter = isMatched) BOOL matched;
 
-- (int)match:(NSArray *)otherCards;
+- (NSInteger)match:(NSArray *)otherCards;
 
 @end
