+++ Matchismo2013Fall/Model/Card.m
@@ -10,7 +10,7 @@
 
 @implementation Card
 
-- (int)match:(NSArray *)otherCards {
+- (NSInteger)match:(NSArray *)otherCards {
     int score = 0;
     
     for (Card *card in otherCards) {
